zamiana int na long, obejście decimal?, metody basket i metoda get product data

+++ SOSM1/SOSM1/Category.cs
@@ -8,8 +8,8 @@ namespace SOSM1
 {
     class Category
     {
-        private int categoryID;
-        public int CategoryID
+        private long categoryID;
+        public long CategoryID
         {
             get
             {
